requirements.txt
Django==1.4.1
South==0.7.6
psycopg2==2.4.5
gunicorn==0.14.3
gevent==0.13.7
Pillow==1.7.7
pylibmc==1.2.3
django-heroku-postgresify==0.2
django-heroku-memcacheify==0.3
